backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
starlette>=0.37.2

# ── ML / Computer Vision ──────────────────────────────────────────────────────
tensorflow>=2.13.0
numpy>=1.24.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0
scikit-learn>=1.4.0

# ── DICOM ─────────────────────────────────────────────────────────────────────
pydicom>=2.4.0

# ── PDF Reports ───────────────────────────────────────────────────────────────
fpdf2>=2.7.9

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.27.0
